--- Organization/prompt/README.md
@@ -1,10 +1,0 @@
-## prompt
-
-**Author:** hkx
-**Version:** 0.0.1
-**Type:** tool
-
-### Description
-
-
-
