--- Organization/prompt/main.py
@@ -1,6 +1,0 @@
-from dify_plugin import Plugin, DifyPluginEnv
-
-plugin = Plugin(DifyPluginEnv(MAX_REQUEST_TIMEOUT=120))
-
-if __name__ == '__main__':
-    plugin.run()
