+++ Organization/prompt/prompt/prompt/requirements.txt
@@ -1,0 +1,2 @@
+dify_plugin>=0.1.0,<0.2.0
+langfuse~=2.60.4
