+++ Organization/prompt/prompt/prompt/tools/prompt.yaml
@@ -1,0 +1,115 @@
+identity:
+  name: prompt
+  author: hkx
+  label:
+    en_US: "Prompt Tool"
+    zh_Hans: "提示词工具"
+
+description:
+  human:
+    en_US: "A tool for generating text from prompts."
+    zh_Hans: "提示词生成工具"
+  llm: "Generates text based on prompts."
+
+parameters:
+  - name: secret_key
+    type: string
+    required: true
+    label:
+      en_US: secret_key
+      zh_Hans: langfuse 秘钥
+    human_description:
+      en_US: Langfuse Key
+      zh_Hans: langfuse 秘钥
+    llm_description: Langfuse Key
+    form: llm
+
+  - name: public_key
+    type: string
+    required: true
+    label:
+      en_US: public_key
+      zh_Hans: langfuse 公钥
+      pt_BR: public_key
+    human_description:
+      en_US: Langfuse public key
+      zh_Hans: langfuse 公钥
+    llm_description: Langfuse public key
+    form: llm
+
+  - name: host
+    type: string
+    required: true
+    label:
+      en_US: host
+      zh_Hans: host
+    human_description:
+      en_US: host
+      zh_Hans: 服务器地址
+    llm_description: host
+    form: llm
+
+  - name: name
+    type: string
+    required: true
+    label:
+      en_US: name
+      zh_Hans: prompt 名称
+      pt_BR: name
+    human_description:
+      en_US: Prompt Name
+      zh_Hans: 可输入Text和Chat类型名称
+    llm_description: Prompt Name
+    form: llm
+
+  - name: label
+    type: string
+    required:
+    label:
+      en_US: label
+      zh_Hans: prompt 标签
+      pt_BR: label
+    human_description:
+      en_US: Prompt label
+      zh_Hans: 置空默认 production, 选填 latest
+    llm_description: Prompt label
+    form: llm
+
+  - name: custom_data
+    type: string
+    required:
+    label:
+      en_US: custom_data
+      zh_Hans: 配置参数
+      pt_BR: custom_data
+    human_description:
+      en_US: custom_data
+      zh_Hans: "配置自定义变量 字典格式，例如: {'key': 'api_key', 'value': 'your_key'}"
+    llm_description: custom_data
+    form: llm
+
+extra:
+  python:
+    source: tools/prompt.py
+
+
+#output_schema:
+#  type: object
+#  properties:
+#    system_value: {type: string}
+#      type: string
+#    user_value: {type: string}
+#  required: [system_value]
+#  required: [system_value, user_value]
+
+output_schema:
+  type: object
+  properties:
+    system_value:
+      type: string
+    user_value:
+      type: string
+    assistant_value:
+      type: string
+    text_value:
+      type: string
