--- Organization/prompt/provider/prompt.py
@@ -1,14 +1,0 @@
-from typing import Any
-
-from dify_plugin import ToolProvider
-from dify_plugin.errors.tool import ToolProviderCredentialValidationError
-
-
-class PromptProvider(ToolProvider):
-    def _validate_credentials(self, credentials: dict[str, Any]) -> None:
-        try:
-            """
-            IMPLEMENT YOUR VALIDATION HERE
-            """
-        except Exception as e:
-            raise ToolProviderCredentialValidationError(str(e))
